undo splitranges
